+++ cg/cli/workflow/base.py
@@ -7,6 +7,7 @@ from .balsamic.base import balsamic as balsamic_cmd
 from .microsalt.base import microsalt
 from .mip_dna.base import mip_dna
 from .mip_rna.base import mip_rna
+from .fluffy.base import fluffy
 
 
 @click.group()
@@ -18,3 +19,4 @@ workflow.add_command(balsamic_cmd)
 workflow.add_command(microsalt)
 workflow.add_command(mip_dna)
 workflow.add_command(mip_rna)
+workflow.add_command(fluffy)