+++ tests/meta/upload/scout/test_meta_upload_scoutapi.py
@@ -3,7 +3,6 @@ from pathlib import Path
 
 import pytest
 import yaml
-
 from cg.meta.upload.scout.mip_config_builder import MipConfigBuilder
 from cg.meta.upload.scout.scout_load_config import MipLoadConfig, ScoutLoadConfig
 from cg.meta.upload.scout.scoutapi import UploadScoutAPI